Add unique serial# to code objects, 64-bit only
This will allow ordering code in a well-defined way, by creation order.
It may not be the optimal order from a locality sense,
but is better for build reproducibility than the current heuristic,
and we don't know that the heurstic improves anything to begin with.

Put ALLOCATE-CODE-OBJECT in 'alloc' where it seems generally happier.

+++ src/runtime/gc-internal.h
@@ -92,9 +92,10 @@ do {                                                                   \
  */
 static inline unsigned int*
 code_fun_table(struct code* code) {
-  return (unsigned int*)((char*)code
-                         + code_header_words(code->header) * N_WORD_BYTES
-                         + fixnum_value(code->code_size) - sizeof (uint16_t));
+  return (unsigned int*)
+      ((char*)code + code_header_words(code->header) * N_WORD_BYTES
+       // Cast out high 32 bits of code_size if lispobj is 64 bits.
+       + fixnum_value((uint32_t)code->code_size) - sizeof (uint16_t));
 }
 static inline unsigned short
 code_n_funs(struct code* code) {
